requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
bottleneck=1.3.4=py39hce1f21e_0
brotlipy=0.7.0=py39h27cfd23_1003
ca-certificates=2022.4.26=h06a4308_0
certifi=2021.10.8=py39h06a4308_2
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39h06a4308_0
cryptography=37.0.1=py39h9ce1e76_0
filelock=3.6.0=pyhd3eb1b0_0
future=0.18.2=py39h06a4308_1
huggingface_hub=0.2.1=pyhd3eb1b0_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py39h06a4308_0
importlib_metadata=4.11.3=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.0=pyhd3eb1b0_0
ld_impl_linux-64=2.38=h1181459_0
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
ncurses=6.3=h7f8727e_2
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numexpr=2.8.1=py39h6abb31d_0
numpy=1.21.5=py39he7a7128_2
numpy-base=1.21.5=py39hf524024_2
openssl=1.1.1o=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py39h295c915_0
pip=21.2.4=py39h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.10.2=cpu_py39hfa7516b_0
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=py39h7f8727e_1
readline=8.1.2=h7f8727e_1
regex=2022.3.15=py39h7f8727e_0
requests=2.27.1=pyhd3eb1b0_0
sacremoses=0.0.43=pyhd3eb1b0_0
scikit-learn=1.0.2=py39h51133e4_1
scipy=1.7.3=py39hc147768_0
setuptools=61.2.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.3=hc218d9a_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h1ccaba5_1
tokenizers=0.11.4=py39h3dcd8bd_1
tqdm=4.64.0=py39h06a4308_0
transformers=4.18.0=py39h06a4308_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39h06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
yaml=0.2.5=h7b6447c_0
zipp=3.8.0=py39h06a4308_0
zlib=1.2.12=h7f8727e_2
